Refactor code Into smaller chunks
