+++ public/lesswrong-editor/src/ckeditor5-footnote/src/footnote.js
@@ -1,5 +1,5 @@
 // @ts-check
-import FootnoteEditing from './footnoteediting';
+import FootnoteEditing from './footnoteEditing/footnoteEditing';
 import FootnoteUI from './footnoteui';
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
